fixed sidebar navigation
Changed two of the sidebar links to the "Calendar" and "MY Tasks" pages. In addition, set the initial state of the navbar to open so that navigating to new pages doesn't abruptly close the navbar. In other words, navbar stays open when navigating to pages, only closes when the user toggles the sidebar.