refactor sidebar toggle button
Originally we had a separate file for holding the sidebar-toggle button, in case there was a case scenario where we used it in more places than just the navbar. However, we realize this is most likely not going to happen and is only causing confusion with 'sidebar.php', the actual sidebar. So the sidebar toggle button file is now merged with the navbar.